Desperately trying to get types to work (Closes #210)
