+++ full-moon-derive/src/lib.rs
@@ -4,7 +4,6 @@ extern crate proc_macro;
 
 mod derive;
 mod node;
-mod symbols;
 mod visit;
 
 use derive::DeriveGenerator;
@@ -19,9 +18,4 @@ pub fn derive_visit(input: TokenStream) -> TokenStream {
 #[proc_macro_derive(Node, attributes(node))]
 pub fn derive_node(input: TokenStream) -> TokenStream {
     node::NodeGenerator::derive(input)
-}
-
-#[proc_macro]
-pub fn symbols(input: TokenStream) -> TokenStream {
-    symbols::parse(input)
 }
